# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ServiceResponse_any-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {},
						"type": "array",
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ServiceResponse_any-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"anyOf": [],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"DocumentCreateRequest": {
				"properties": {
					"projectId": {
						"type": "string"
					},
					"documentName": {
						"type": "string"
					},
					"size": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"type": {
						"type": "string",
						"nullable": true
					},
					"addedAt": {
						"type": "string",
						"format": "date-time"
					},
					"documentUrl": {
						"type": "string",
						"nullable": true
					},
					"userId": {
						"type": "string"
					},
					"publicId": {
						"type": "string",
						"nullable": true
					},
					"deleteToken": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"documentName",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DocumentUpdateRequest": {
				"properties": {
					"projectId": {
						"type": "string"
					},
					"documentName": {
						"type": "string"
					},
					"size": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "string"
					},
					"addedAt": {
						"type": "string",
						"format": "date-time"
					},
					"documentUrl": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"publicId": {
						"type": "string"
					},
					"deleteToken": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"optionalJwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "community-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Ikundabayo placide"
		}
	},
	"paths": {
...
		"/api/documents": {
			"get": {
				"operationId": "GetDocuments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Documents"
				],
				"security": [
					{
						"jwt": [
							"project:read"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "projectId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateDocument",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Documents"
				],
				"security": [
					{
						"jwt": [
							"project:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DocumentCreateRequest"
							}
						}
					}
				}
			}
		},
		"/api/documents/{documentId}": {
			"get": {
				"operationId": "GetDocumentById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					},
					"404": {
						"description": "Document not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Documents"
				],
				"security": [
					{
						"jwt": [
							"project:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "documentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateDocument",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Documents"
				],
				"security": [
					{
						"jwt": [
							"project:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "documentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DocumentUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteDocument",
				"responses": {
					"204": {
						"description": "No Content",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Documents"
				],
				"security": [
					{
						"jwt": [
							"project:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "documentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/documents/{documentId}/blob": {
			"get": {
				"operationId": "GetFileBlob",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string",
									"format": "byte"
								}
							}
						}
					},
					"400": {
						"description": "Document URL not available",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					},
					"404": {
						"description": "Document not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Documents"
				],
				"security": [
					{
						"jwt": [
							"project:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "documentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}